New test: Don't allow multiple authentication methods
As found when integrating with my Micropub server, using an OAuth2
compliant resource server configuration, not all clients are
well-formed, and may send multiple instances of the access token, which
is incorrect according to OAuth2.

We should add a test to capture the fact that this should not be
possible, for both a client, and a server.

This requires we:

- add a new server test, and hook it in
- add a new client features for authentication tests for multipart form
  requests
- update the wording of the existing client authentication tests to make
  sure they're clearly for `x-www-form-urlencoded` requests
- add our client test to the implementation report view
- add the validation for the new client tests, and tag them against the
  correct features

Closes #84.